[PDI-13408] - Set AutomaticXmlMetaStoreCache as default cache. + Add Guava lib rev - 17. + Added variables for revisions version.
