+++ build.properties
@@ -6,3 +6,9 @@ impl.title=Metastore
 impl.productID=metastore
 javadoc.packagenames=org.pentaho.*
 
+# Dependency revisions
+dependency.commons-codec.revision=1.5
+dependency.guava.revision=17.0
+dependency.junit.revision=4.7
+dependency.mockito-all.revision=1.8.4
+
