+++ ivy.xml
@@ -17,8 +17,9 @@
     
   <dependencies defaultconf="default->default">
 
-    <dependency org="commons-codec"       name="commons-codec"       rev="1.5"              transitive="false"/>
-    <dependency org="junit"               name="junit"               rev="4.7"              transitive="false" conf="test->default"/>
-    <dependency org="org.mockito"         name="mockito-all"         rev="1.8.4"            transitive="false" conf="test->default"/>          
+    <dependency org="commons-codec"       name="commons-codec"       rev="${dependency.commons-codec.revision}"      transitive="false"/>
+    <dependency org="com.google.guava"    name="guava"               rev="${dependency.guava.revision}"              transitive="false"/>
+    <dependency org="junit"               name="junit"               rev="${dependency.junit.revision}"              transitive="false" conf="test->default"/>
+    <dependency org="org.mockito"         name="mockito-all"         rev="${dependency.mockito-all.revision}"        transitive="false" conf="test->default"/>
   </dependencies>
 </ivy-module>
